Removed the usage of the memory pool for gRPC transport. The buffer is now pooled within gRPC-Go. 



RELEASE NOTES:
 - Added a benchmark with different scenarios to verify that the performance is not impacted.
 
-------
Results of running the new benchmark before and after the changes:
Before
➜   go test -bench=BenchmarkIntegrationGRPCMemorypool -benchmem ./encoding/protobuf/testing/...
goos: darwin
goarch: arm64
pkg: go.uber.org/yarpc/encoding/protobuf/testing
cpu: Apple M2 Max
BenchmarkIntegrationGRPCMemorypool/SmallOnly-12                    20000             73996 ns/op             13114 B/op_sys           12691 B/op_total      12690 B/op        211 allocs/op
BenchmarkIntegrationGRPCMemorypool/LargeOnly-12                    20000           1155262 ns/op           3169364 B/op_sys        3177527 B/op_total    3177527 B/op        279 allocs/op
BenchmarkIntegrationGRPCMemorypool/Mixed95_5-12                    20000            123958 ns/op                36.04 B/op_sys      171401 B/op_total     171401 B/op        214 allocs/op
PASS
ok      go.uber.org/yarpc/encoding/protobuf/testing     28.282s

After
➜   go test -bench=BenchmarkIntegrationGRPCMemorypool -benchmem ./encoding/protobuf/testing/...
goos: darwin
goarch: arm64
pkg: go.uber.org/yarpc/encoding/protobuf/testing
cpu: Apple M2 Max
BenchmarkIntegrationGRPCMemorypool/SmallOnly-12                    20000             74929 ns/op             13167 B/op_sys          12691 B/op_total      12690 B/op        211 allocs/op
BenchmarkIntegrationGRPCMemorypool/LargeOnly-12                    20000           1123882 ns/op           3169131 B/op_sys        3177348 B/op_total    3177347 B/op        268 allocs/op
BenchmarkIntegrationGRPCMemorypool/Mixed95_5-12                    20000            122972 ns/op                36.04 B/op_sys      171338 B/op_total     171338 B/op        213 allocs/op
PASS
ok      go.uber.org/yarpc/encoding/protobuf/testing     27.698s


SmallOnly:
Before: 73,996 ns/op, 12,691 B/op, 211 allocs/op
After: 74,929 ns/op, 12,690 B/op, 211 allocs/op
Change: Slightly slower (+1.3%), virtually identical memory usage
LargeOnly:
Before: 1,155,262 ns/op, 3,177,527 B/op, 279 allocs/op
After: 1,123,882 ns/op, 3,177,347 B/op, 268 allocs/op
Change: Slightly faster (-2.7%), slightly fewer allocations (-11 allocs)
Mixed95_5:
Before: 123,958 ns/op, 171,401 B/op, 214 allocs/op
After: 122,972 ns/op, 171,338 B/op, 213 allocs/op
Change: Slightly faster (-0.8%), virtually identical memory usage