@pulkit4tech 2025-05-15
This is great!

let's add mallocs as well and also comparison for (99% small - 1% large) & (90%small && 10%large)